Move ChainState and ChainStateDelta to chain-core
