+++ chain-storage/src/lib.rs
@@ -1,6 +1,5 @@
 extern crate chain_core;
 
 pub mod error;
-pub mod chain;
 pub mod store;
 pub mod memory;
